backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
PyJWT
python-dotenv
psycopg2-binary
bcrypt
